worker/requirements.txt
atomicwrites==1.1.5
attrs==18.1.0
autopep8==1.3.5
certifi==2018.4.16
click==6.7
future==0.16.0
more-itertools==4.3.0
pg8000==1.12.2
pluggy==0.7.1
psycopg2-binary==2.7.5
py==1.5.4
pycodestyle==2.4.0
pytest==3.7.1
python-dotenv==0.9.1
python-telegram-bot==10.1.0
redis==2.10.6
rq==0.12.0
six==1.11.0
SQLAlchemy==1.2.10
